Slightly tweaked how links work
